Refactor: Relocate `DrawingBar.kt` and remove `DevicePreviews.kt`
This commit moves the `DrawingBar.kt` file from the `com.mshdabiola.drawing` package to `com.mshdabiola.ui`. It also deletes the `DevicePreviews.kt` file.

**Key Changes:**

*   **File Move:**
    *   `core/ui/src/commonMain/kotlin/com/mshdabiola/drawing/DrawingBar.kt` moved to `core/ui/src/commonMain/kotlin/com/mshdabiola/ui/DrawingBar.kt`.
    *   The package declaration in `DrawingBar.kt` has been updated accordingly.
*   **File Deletion:**
    *   `core/ui/src/commonMain/kotlin/com/mshdabiola/ui/DevicePreviews.kt` has been removed.
*   **Import Updates:**
    *   Unused imports (`painterResource`, `stringResource`, `Preview`, `NoteTab`, `NoteTabRow`, `NoteTextButton`, `FlowLayout2`) were removed from `DrawingBar.kt`.

This refactoring aims to improve the organization of UI components.
