--- core/ui/src/commonMain/kotlin/com/mshdabiola/ui/DevicePreviews.kt
@@ -1,17 +1,0 @@
-/*
- *abiola 2024
- */
-
-package com.mshdabiola.ui
-
-import androidx.compose.ui.tooling.preview.Preview
-
-/**
- * Multipreview annotation that represents various device sizes. Add this annotation to a composable
- * to render various devices.
- */
-@Preview(name = "phone", device = "spec:shape=Normal,width=360,height=640,unit=dp,dpi=480")
-@Preview(name = "landscape", device = "spec:shape=Normal,width=640,height=360,unit=dp,dpi=480")
-@Preview(name = "foldable", device = "spec:shape=Normal,width=673,height=841,unit=dp,dpi=480")
-@Preview(name = "tablet", device = "spec:shape=Normal,width=1280,height=800,unit=dp,dpi=480")
-annotation class DevicePreviews
