+++ core/ui/src/commonMain/kotlin/com/mshdabiola/ui/DrawingBar.kt
@@ -1,4 +1,4 @@
-package com.mshdabiola.drawing
+package com.mshdabiola.ui
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -33,15 +33,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.mshdabiola.designsystem.component.NoteTab
-import com.mshdabiola.designsystem.component.NoteTabRow
-import com.mshdabiola.designsystem.component.NoteTextButton
 import com.mshdabiola.model.DRAW_MODE
-import com.mshdabiola.ui.FlowLayout2
 import kotlinx.coroutines.launch
 import com.mshdabiola.designsystem.R as Rd
